Feat: Add neighbour coordinates function
Without normalization if we need to check whether a cell is going to be
alive or not, we need to potentially check all around the elements
whether in next iteration of the game of life they are going to be alive
or not.
This function creates the neighbours coordinates without normalization.